Began work on the logo

+++ inc/col.h
@@ -7,6 +7,6 @@
 #define COL_FONT_SEL 128
 
 void col_init(void);
-void col_puts(u8 x, u8 y, char *s);
+void col_puts(u16 x, u16 y, char *s);
 
 #endif
